Refactor code to additional classes (Ui, Storage, Parser and TaskList)
